Backend/requirements.txt
fastapi
uvicorn
fastapi-login
python-multipart
bcrypt
exponent_server_sdk
python-dotenv